ai_solution_architect_v2/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
pdfplumber
python-docx
pydantic>=2.9